refactor: printWinner파일 에서 getWinner 함수를 다른 파일로 분리
